% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimateParameters.R
\name{estimateParameters}
\alias{estimateParameters}
\title{Estimate parameters in SDE model.}
\usage{
estimateParameters(
  obsdata,
  inputdata,
  parameters,
  parameter_map = NULL,
  x0 = 3,
  p0 = 0.1 * diag(1),
  ode.steps = 10,
  ode.dt = diff(inputdata$t)/ode.steps,
  ode.n = rep(ode.steps, nrow(inputdata) - 1),
  lb = -Inf,
  ub = Inf,
  silent = TRUE,
  trace = 0,
  control = list(trace = trace, eval.max = 10000, iter.max = 10000),
  hessian = TRUE
)
}
\arguments{
\item{obsdata}{description}

\item{inputdata}{description}

\item{parameters}{description}

\item{lb}{description}

\item{ub}{description}

\item{silent}{description}

\item{hessian}{description}

\item{hyperdata}{description}
}
\description{
Maximum likelihood based estimation of state space parameters.
}
