% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makePredictionUsingR.R
\name{makePredictionUsingR}
\alias{makePredictionUsingR}
\title{Perform state predictions from a specified model}
\usage{
makePredictionUsingR(
  inputdata,
  hyperdata = list(ode.timestep.size = diff(inputdata$t), ode.timesteps = rep(1,
    nrow(inputdata) - 1)),
  parvec = c(theta = 10, mu = 1, sigma_x = 1, sigma_y = 0.1),
  X = 1,
  P = 0.1
)
}
\arguments{
\item{inputdata}{description}

\item{hyperdata}{description}

\item{parameters}{description}
}
\description{
Calculates predictions
}
